% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2.R
\name{pr2_database}
\alias{pr2_database}
\title{PR2 database - Eukaryotic 18S rRNA sequences.}
\usage{
pr2_database()
}
\value{
A data frame
}
\description{
The PR2 database is provided as a data frame.

This is a join between the following tables:
\itemize{
       \item{pr2_main}
       \item{pr2_taxonomy}
       \item{pr2_sequence}
       \item{pr2_metadata}
       \item{pr2_traits}
       \item{pr2_silva}
       \item{euribo}
       \item{taxonomy_gbif}
       \item{taxonomy_worms}
  }
The metadata contains different types of fields
\itemize{
       \item{gb_ : originating from the GenBank entry}
       \item{eukref_ : annotated by the Eukref project}
       \item{pr2_ : annotated by pr2 such latitude and longitude}
       \item{eukribo_: information from the EukRibo database}
  }
Description of fields: \url{https://pr2-database.org/documentation/pr2-fields}
}
\examples{

# Read the whole database
my_pr2 <- pr2_database()

# Select a specific genus
pr2_ostreo <- pr2_database()
pr2_ostreo <- dplyr::filter(pr2_ostreo , genus == "Ostreococcus")
pr2_ostreo <-  dplyr::select(pr2_ostreo, pr2_accession, species)
head(pr2_ostreo)
}
