% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2.R
\name{pr2_taxonomy}
\alias{pr2_taxonomy}
\title{PR2 database - Taxonomy.}
\usage{
pr2_taxonomy()
}
\value{
A data frame
}
\description{
The PR2 database taxonomy is provided as a data frame with the number of sequences for each taxon
}
\examples{

# Read the whole database taxonomy
my_pr2_taxo <- pr2_taxonomy()

# Select a specific genus
my_pr2_taxo <- dplyr::filter( my_pr2_taxo , genus == "Ostreococcus")
head(my_pr2_taxo)
}
