% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pr2}
\alias{pr2}
\title{PR2 database - Eukaryotic 18S rRNA sequences.}
\format{
A data frame
}
\source{
PR2 database - \url{https://pr2-database.org/}
}
\usage{
pr2
}
\description{
The PR2 database is provided as a data frame.
}
\details{
This is a join between the following tables:
\itemize{
       \item{pr2_main}
       \item{pr2_taxonomy}
       \item{pr2_sequence}
       \item{pr2_metadata}
       \item{pr2_silva}
  }
The metadata contains three types of fields
\itemize{
       \item{gb_ : originating from the GenBank entry}
       \item{eukref_ : annotated by the Eukref project}
       \item{pr2_ : annotated by pr2 such latitude and longitude}
  }
Description of fields: \url{https://pr2-database.org/documentation/pr2-fields}
}
\keyword{datasets}
